requirements.txt
.dependencies/tesserocr-2.6.0-cp38-cp38-win_amd64.whl
keyboard==0.13.5
mouse==0.7.1
numpy==1.24.4
opencv-python==4.9.0.80
pillow==10.2.0
playsound==1.2.2
pyautogui==0.9.54
pygetwindow==0.0.9
